Modularize server and secure configs
