+++ package.json
@@ -13,17 +13,14 @@
     "@aws-sdk/client-s3": "^3.666.0",
     "@aws-sdk/s3-request-presigner": "^3.666.0",
     "@googlemaps/google-maps-services-js": "^3.4.0",
-    "aws-sdk": "^2.1644.0",
     "bcryptjs": "^2.4.3",
-    "correios-brasil": "^3.0.6",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
     "firebase": "^10.12.2",
     "nodemailer": "^6.9.16",
     "pluralize": "^8.0.0",
-    "stripe": "^16.8.0",
-    "twilio": "^5.3.6"
+    "stripe": "^16.8.0"
   },
   "devDependencies": {
     "nodemon": "^3.1.2"
